build: fixes for building for (non-Android-)Linux
This is a preparatory step for adding build scripts for openmobilemaps
for (non-Android-)Linux.

Fix various missing includes and other small errors that are accepted by
the Android/iOS build toolchains, but are errors when attempting to
build with a generic GCC/Clang.
In particular, change `abs` to `std::abs` when used with float arguments
as suggested by clang (-Wabsolute-value); this avoids the implicit and
likely undesired truncation of floats with `abs`.
Change `#ifdef __ANDROID__` to `#ifdef __linux__` as OpenGL feature
toggle for now.

Fix misc small warnings, making this build cleanly (on GCC 13.2, clang
18.1) with "only" disabling "-Wno-reorder -Wno-deprecated" (note that's
with default warnings, without -Wall).