Java bindings: add BaseDataLoader, OffscreenMapRenderer quirks
Add abstract BaseDataLoader to share the common logic for DataLoaders in
Java and use this in the HttpDataLoader.
Remove the LocalDataLoader as it's never been used and seems redundant
with the LocalDataProvider-option on the Tiled2dMapVectorLayerBuilder.
With the BaseDataLoader, loading from resources can be easily added in
project-specific subclasses.

Extend the "quirks" (draw the frame repeatedly) to make the
OffscreenMapRenderer work reliably again after rebasing these java
bindings.