Fix memory leak from gpc_polygon in TileWrapper
TileWrapper keeps a gpc_polygon which is not freed on destruction of the TileWrapper.
In some cases this was explicitly freed, but not everywhere.
To simplify this, wrap the gpc_polygon in a semi-safe holder to ensure
its always destroyed.
Unfortunately the gpc_... API does not use const pointers for the input
arguments, so the wrapper cannot be made properly safe.

Also, add missing intitialization for Tiled2dMapSource::curT/curZoom,
reported by valgrind.