Java bindings: handle BaseDataLoader futures are available immediately
In the BaseDataLoader::loadData/TextureAsync, a child class can return a
future that may already be completed (or failed) at the time that it is being
created. For such a future, the whenComplete callback is called
immediately from the same callstack.
During creating an entry for the ongoing load via
ConcurrentHashMap::compute, this whenComplete callback would then try to
remove that same entry. This is both nonsensical and illegal.
Use whenCompleteAsync to ensure that this is deferred until after the
entry has been created.