Prioritize data loads per loader, so fallback cannot block primary
In 5d95aef0f42860e5162477926040cab54f764dc5, a limit to the number of
concurrent tile loads was introduced.
As a fallback to the next loader in the loader cascade was performed
immediately, slow "fallback" loaders could block the subsequent tile
loads from the "primary" loaders.
In the somewhat common loader cascade setup to load from offline data
first and then fallback to a remote request, blocking the local loads
with a slow remote loader can be disastrous.

For now, we keep the same fixed number of concurrent load "slots" and
strictly prioritize by loader index. That is, all loads for loader 0 are
processed before the loads for loader 1, etc.
Alternatively, the slots could be divided in any scheme among the
loaders, as long as the primary loader is never starved.

The Tiled2dMapSource now has to explicitly handle a NOP from the last
loader, and the case that no loaders are configured; this is now treated
like "not found".

Related to this, also, is that Tiled2dVectorGeoJsonSource now sets
loaderCount to exactly 1, via an unused dummy loader entry in the
initialization of its base class.