github: fixup XML postprocessing of unit test results
Cursed XML postprocessing failed due to a '>' character in a test name.
This is not consistently escaped by the catch2 XML reporter.

Aside; the catch2 tests also include benchmark tests, which take a
relatively long time to run. As we ignore the measurements here, make
the benchmark tests run through as quickly as possible.