Improved label anti-aliasing for small font sizes
Using `fwidth(sdf)` to determine the sdf-to-screen-pixel-ratio,
as the input for the smoothstep used for anti-aliasing works reasonably
well as long as the underlying distance function is sdf smooth.
It is however very ineffective with small features (such as fonts at
small sizes). Neighboring pixels can be at the opposite sides of a
feature and the observed derivates (fwidth) vanish. This leads to a
under-estimation of the sdf-to-screen-pixel-ratio. We would then use a
small blur radius for pixels around thin font features, leading to
noticable, hard black/white pixels.

Instead, we now use the derivative of the texture coordinates to
determine the sdf-to-screen-pixel-ratio, which is always smooth.
This approach needs as an additional input the range of pixel-distances
represented by the SDF. This information is available as metadata in the
font manifest.

Alternatively, we could do away with the derivatives alltogether, if we
could determine the intended text size (in pixel size) in the shader.
Then the sdf-to-screen-pixel-ratio could be determined simply with
`textSize/referenceFontSize * distanceRange`, where referenceFontSize is
available as `info.size` in the font-manifest.
The intended text size is, however, not currently available in the
shader. Passing this as additional information for each glyph seems
excessively redundant.

Aside: the reference font size from the font manifest used to be
multiplied by the device pixel density factor (device PPI / 160). This
seemed unnecessarily confusing and was moved to the only place that
still seemed to rely on this behaviour (TextLayerObject). This would
only have been relevant for the alternative approach outlined above, and
thus is not strictly necessary for the functional changes in this
changeset.