Interned strings for FeatureContext and style expressions
Avoid frequent string comparison during style expression evaluation by
interning strings and only using a lightweight reference/handle in place
of the string keys.

To avoid potentially leaking memory forever, the StringInterner table
is not global but scoped to a vector layer.

At the system border, we need sometimes need to get the strings back,
in particular for FeatureContext::getFeatureInfo.
For this, we need to keep some additional (weak-)references to the
vector layer containing the StringInterner table.