Enable -Wunused and fix warnings
In CMakeLists for android and JVM bindings, enable -Wunused and fix the
resulting warnings in our code. This does not enable -Wunused-parameters;
enabling this additionally leads to too many warnings right now.
Add djinni and other external library headers as system headers to
suppress warnings in there. Unfortunately, clang thinks that some
deprecation warnings need to be shown even in system headers, so we
have to disable these globally.

- projectOntoUnitSphere was kept as an unused member in many opengl
  shaders
- persistingSymbolPlacement was apparently piped through to SymbolGroup
  and SymbolObject where it was not used
- a few non-trivial code snippets resulting in unused variables
  commented out instead of removing them.