github: fix JVM release builds, use same runner for jvm_build and jvm_package_deploy
Release builds were failing because they were running on the older
ubuntu github runners. This was different to the runners that was
running the test build, which meant that some build problems (with older
C++ library) only showed up during release build.

The reason to use the older runners was to ensure that we link against
the ubuntu-22.04 glibc. This is no longer needed for our internal builds.

Also, expand the jvm_build job instead of using the workflow template
(to set runs-on). This makes it easy to enable ccache.